feat!: command executors and tab completers registered through annotations
Classes annotated with me.ancliz.minecraft.annotations.CommandExecutor
or me.ancliz.minecraft.annotations.TabCompleter will be automatically
registed to the plugin for the specified command in their paramter list.
Plans to make this an array value rather than a single command are in
place, as well as building the commands section in the plugin.yml from
the annotated classes.

Classes annotated with the above must be subtypes of
me.ancliz.minecraft.commands.DefaultCommandExecutor and
me.ancliz.minecraft.commands.DefaultTabCompleter respectively.

BREAKING CHANGE: implementation of how executors and completers are
registered has changed and will need to be updated accordingly